+++ src/Enums/DataTarget.php
@@ -10,5 +10,6 @@ enum DataTarget: string
     case OneOffQuote = 'OneOffQuote';
     case Receiveable = 'AccountingInvoice';
     case Organization = 'Organization';
+    case Container = 'Container';
     case Staff = 'Staff';
 }