+++ src/Requests/NativeContainerRetrieval.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Oliverbj\Cord\Requests;
+
+class NativeContainerRetrieval extends NativeRequest
+{
+    public function schema(): array
+    {
+        return $this->defineSchema($this->cord->criteriaGroups);
+    }
+
+    private function defineSchema(array $criteriaGroups): array
+    {
+        return [
+            'Body' => [
+                'ContainerType' => $criteriaGroups,
+            ],
+        ];
+    }
+}